requirements.txt
bokeh==1.0.2
numpy==1.15.1
PyGithub==1.43.3
